Made configuration file installer
It's not super super flexible -- just enough to "work for me" :P

So you can install to PREFIX, so long as PREFIX is

 - /usr/local
 - /usr
 - $HOME/.local

Anything else will not work.

By default it will install to /usr/local.

This will fetch config files from the folders /etc/em or ~/.config/em;
use these if you wish to override the settings in /usr/local/share/em,
/usr/share/em, or ~/.local/share/em.