Switched to Tk rather than Qt
Motivation was the scary huge size of the venvs created by pulling in
Qt (~500MB) which is ridiculous for a small tool.

Had to cruft something together for Tk, but hey!

BTW, I didn't use PySimpleGUI because there was a noticeable lag when
it was slurping the GIF into memory.  Figured it was better to go
closer to the metal here.