Bump version number to 0.3.0
Switched from PySide to tkinter!  It's a little slower, but switching
took venv sizes from 470M to 8M.  Yeaaaaahhhhhhhhhhhhh, kinda hard to
claim this is a "small tool" when it pulls in a half-gig of
unnecessary junk.

Otherwise, there's very little difference from 0.2.0 in look & feel,
and no difference in functionality.